# openapi.yaml
openapi: 3.1.0
info:
  title: Parento Backend API
  version: 1.0.0
  description: |
    API foundation for Parento. This document includes administrator authentication,
    Phase 5 enrollment, and the Phase 6.1 secure device communication/command foundation.
    Phase 6.1 exposes only the neutral FUTURE_COMMAND envelope; no device-control capability is implemented.
servers:
  - url: /
paths:
...
  /api/v1/devices/{deviceId}/commands:
    post:
      summary: Create an allowlisted command envelope for an owned device
      operationId: createDeviceCommand
      security: [{ bearerAuth: [] }]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, version, payload]
              properties:
                type: { type: string, enum: [FUTURE_COMMAND] }
                version: { type: integer, enum: [1] }
                payload:
                  {
                    type: object,
                    additionalProperties: false,
                    maxProperties: 0,
                  }
                idempotencyKey: { type: string, maxLength: 128, nullable: true }
                correlationId: { type: string, maxLength: 128, nullable: true }
      responses:
        '201': { description: Command created }
        '200':
          { description: Existing command returned for an idempotent retry }
        '400': { $ref: '#/components/responses/ApiErrorResponse' }
        '403': { $ref: '#/components/responses/ApiErrorResponse' }
        '404': { $ref: '#/components/responses/ApiErrorResponse' }
        '409': { $ref: '#/components/responses/ApiErrorResponse' }
  /api/v1/devices/{deviceId}/commands/{commandId}:
    get:
      summary: Get an owned command
      operationId: getDeviceCommand
      security: [{ bearerAuth: [] }]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: commandId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Command status }
        '401': { $ref: '#/components/responses/ApiErrorResponse' }
        '403': { $ref: '#/components/responses/ApiErrorResponse' }
        '404': { $ref: '#/components/responses/ApiErrorResponse' }
  /api/v1/devices/{deviceId}/commands/{commandId}/cancel:
    post:
      summary: Cancel a non-terminal command
      operationId: cancelDeviceCommand
      security: [{ bearerAuth: [] }]
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: commandId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Cancelled command }
        '409': { $ref: '#/components/responses/ApiErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    deviceSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - AUTHENTICATION_REQUIRED
                - INVALID_REQUEST
                - INVALID_CREDENTIALS
                - AUTHORIZATION_DENIED
                - RESOURCE_NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - METHOD_NOT_ALLOWED
                - REQUEST_TOO_LARGE
                - INTERNAL_SERVER_ERROR
                - SERVICE_UNAVAILABLE
                - DATABASE_UNAVAILABLE
                - NOT_FOUND
                - ENROLLMENT_NOT_FOUND
                - ENROLLMENT_EXPIRED
                - ENROLLMENT_ALREADY_CONSUMED
                - ENROLLMENT_STATE_CONFLICT
                - INVALID_ENROLLMENT_VERIFICATION
                - DEVICE_IDENTITY_CONFLICT
                - DEVICE_AUTHENTICATION_REQUIRED
                - DEVICE_AUTHORIZATION_DENIED
                - DEVICE_SESSION_INVALID
                - DEVICE_NOT_READY
                - DEVICE_NOT_FOUND
                - COMMAND_NOT_FOUND
                - COMMAND_STATE_CONFLICT
                - COMMAND_IDEMPOTENCY_CONFLICT
                - COMMAND_PAYLOAD_TOO_LARGE
                - UNSUPPORTED_COMMAND_TYPE
            message:
              type: string
            metadata:
              type: object
        requestId:
          type: string
...
  responses:
    ApiErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...